+++ api/lib/routes.js
@@ -27,6 +27,7 @@ import * as organizationLearners from './application/organization-learners/index
 import * as organizations from './application/organizations/index.js';
 import * as organizationsAdministration from './application/organizations-administration/index.js';
 import * as passwords from './application/passwords/index.js';
+import * as poleEmploi from './application/pole-emploi/index.js';
 import * as scoOrganizationLearners from './application/sco-organization-learners/index.js';
 import * as sessions from './application/sessions/index.js';
 import * as supOrganizationLearners from './application/sup-organization-learners/index.js';
@@ -63,6 +64,7 @@ const routes = [
   organizations,
   organizationsAdministration,
   passwords,
+  poleEmploi,
   organizationInvitations,
   scoOrganizationLearners,
   supOrganizationLearners,
