+++ api/tests/integration/infrastructure/repositories/pole-emploi-sending-repository_test.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
-import { config as settings } from '../../../../../../lib/config.js';
-import * as poleEmploiSendingRepository from '../../../../../../src/prescription/campaign-participation/infrastructure/repositories/pole-emploi-sending-repository.js';
-import { databaseBuilder, domainBuilder, expect, knex } from '../../../../../test-helper.js';
+import { config as settings } from '../../../../lib/config.js';
+import * as poleEmploiSendingRepository from '../../../../lib/infrastructure/repositories/pole-emploi-sending-repository.js';
+import { databaseBuilder, domainBuilder, expect, knex } from '../../../test-helper.js';
 const poleEmploiSendingFactory = databaseBuilder.factory.poleEmploiSendingFactory;
 
 describe('Integration | Repository | PoleEmploiSending', function () {
