+++ api/tests/unit/application/pole-emploi/pole-emploi-controller_test.js
@@ -1,7 +1,7 @@
-import { poleEmploiController } from '../../../../../src/prescription/campaign-participation/application/pole-emploi-controller.js';
-import { usecases } from '../../../../../src/prescription/campaign-participation/domain/usecases/index.js';
-import { UnprocessableEntityError } from '../../../../../src/shared/application/http-errors.js';
-import { catchErr, expect, hFake, sinon } from '../../../../test-helper.js';
+import { UnprocessableEntityError } from '../../../../lib/application/http-errors.js';
+import { poleEmploiController } from '../../../../lib/application/pole-emploi/pole-emploi-controller.js';
+import { usecases } from '../../../../lib/domain/usecases/index.js';
+import { catchErr, expect, hFake, sinon } from '../../../test-helper.js';
 
 describe('Unit | Controller | pole-emploi-controller', function () {
   describe('#getSendings', function () {
@@ -40,6 +40,7 @@ describe('Unit | Controller | pole-emploi-controller', function () {
         poleEmploiService.decodeCursor.withArgs(badCurseur).rejects(new SyntaxError());
         // when
         const error = await catchErr(poleEmploiController.getSendings)(request, hFake, { poleEmploiService });
+
         //then
         expect(error).to.be.an.instanceof(UnprocessableEntityError);
         expect(error.message).to.equal('The provided cursor is unreadable');
